Make tests mirror src layout and test-file naming
Tests are meant to mirror src/, but several files had drifted:

- ecmwf/ifs_ens/forecast_46_day_dynamical_dataset_test.py sat at the
  ifs_ens level while its module lives in forecast_46_day_1_5_degree/.
- contrib/uarizona/swann/ held the tests for the swann/analysis/ package.
- deploy_test.py existed both at the tests root (covering
  common/deploy.py) and under tests/common/; merge them into the mirrored
  location.
- Twenty files used the test_<module>.py order against 128 using
  <module>_test.py, which is also what initialize-new-integration
  scaffolds; rename them to the majority convention.

Record the convention in AGENTS.md so the layout stays derivable from src.
Test collection is unchanged at 2430.
