+++ AGENTS.md
@@ -39,7 +39,7 @@ src/reformatters/
 ├── example_materialized/    # Teaching template for a materialized (time-optimized) dataset
 └── example_virtual/         # Teaching template for a virtual (spatial-optimized) dataset
 
-tests/                       # Mirrors src/ structure
+tests/                       # Mirrors src/ structure; `<module>.py` is tested by `<module>_test.py` alongside it
 docs/
 ├── dataset_development_guide.md      # explore→implement→backfill→validate→publish
 ├── implementation_guide.md           # Step-by-step new dataset implementation walkthrough
@@ -177,7 +177,7 @@ See [docs/parallel_processing.md](docs/parallel_processing.md) for details on co
 * Format: `uv run ruff format`
 * Test: `uv run pytest`
 * Fast/unit tests: `uv run pytest -m "not slow"`
-* Single test: `uv run pytest tests/test_file.py::test_function_name`
+* Single test: `uv run pytest tests/path/to/module_test.py::test_function_name`
 * Important: always run all of these checks before committing, do not skip them: `uv run ruff format && uv run ruff check --fix && uv run ty check`.
 * Run the generate-manual-workflows prek step when adding or removing a dataset.
 * Use `uv run ...` to run python commands in the environment, e.g. `uv run python -c "..."`, `uv run src/scripts/foo.py`. Do not call `python3` when working in this repo.
