+++ tests/common/binary_rounding_test.py
